Move out ws tests to regress_ws.c
